fix app args
