+++ evalscope/report/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
 
+from evalscope.report.app_arguments import add_argument
 from evalscope.report.combinator import gen_table, get_data_frame, get_report_list
 from evalscope.report.generator import ReportGenerator
 from evalscope.report.utils import Category, Report, ReportKey, Subset
